RAG_Chatbot/requirements.txt
streamlit
langchain
pytesseract
pdf2image
PyMuPDF
together
faiss-cpu
sentence-transformers
Pillow
